Nova vers├úo: 1.0.5

+++ update.bat
@@ -1,5 +1,17 @@
 @echo off
 cd /d %~dp0
-git fetch --all
-git reset --hard origin/main
+
+:: Verifica se o diretório atual é um repositório Git
+if not exist ".git" (
+    echo Repositório Git não encontrado.
+    echo Clonando o repositório do GitHub...
+    git clone https://github.com/mathfemar/Antera---AIA.git .
+) else (
+    echo Repositório Git encontrado.
+    echo Sincronizando com o repositório remoto...
+    git fetch --all
+    git reset --hard origin/main
+    git clean -fdx
+)
+
 exit
